[1/3] Rename shared.py to es_test_case.py
